@@ -7,52 +7,158 @@ from frappe.model.document import Document
 from frappe import _
 from frappe.utils import getdate, cint, now, today, flt
 from erpnext.accounts.utils import get_fiscal_year
-from frappe.utils.password import set_encrypted_password, get_decrypted_password
 import re
 
 
 class PortugalDocumentSeries(Document):
 	def validate(self):
-		"""Validações do documento"""
+		"""
+		✅ ADAPTADO: Validações alinhadas com nova abordagem
+		"""
 		self.validate_series_name()
-		self.validate_prefix_format()
+		self.validate_prefix_format_new()
 		self.validate_document_type()
 		self.check_duplicate_series()
 		self.validate_sequence_number()
 		self.validate_company_compliance()
-		self.validate_prefix_for_document_type()
-		self.validate_series_configuration_link()
+		self.validate_prefix_for_document_type_new()
+		self.sync_with_portugal_series_configuration()
 
-	def validate_series_configuration_link(self):
-		"""Validar ligação com Portugal Series Configuration"""
-		# ✅ CORREÇÃO: Verificar se campo existe antes de acessar
-		series_config = getattr(self, 'series_configuration', None)
+	def sync_with_portugal_series_configuration(self):
+		"""
+		✅ NOVO: Sincronizar com Portugal Series Configuration
+		"""
+		try:
+			# Buscar configuração correspondente
+			config = frappe.db.get_value("Portugal Series Configuration", {
+				"prefix": self.prefix,
+				"company": self.company,
+				"document_type": self.document_type
+			}, ["name", "validation_code", "is_communicated"], as_dict=True)
 
-		if series_config:
-			try:
-				# Verificar se a configuração existe e está comunicada
-				config = frappe.get_doc("Portugal Series Configuration", series_config)
-
-				if not config.is_communicated:
-					frappe.throw(
-						_("A configuração da série '{0}' ainda não foi comunicada à AT").format(
-							config.series_name))
-
-				if not config.validation_code:
-					frappe.throw(
-						_("A configuração da série '{0}' não tem código ATCUD válido").format(
-							config.series_name))
-
-				# Sincronizar dados da configuração
+			if config:
+				# Sincronizar dados
 				self.validation_code = config.validation_code
-				self.at_environment = config.at_environment
-				self.communication_date = config.communication_date
+				self.is_communicated = config.is_communicated
+				self.series_configuration = config.name
 
-			except frappe.DoesNotExistError:
-				frappe.throw(_("Configuração de série '{0}' não encontrada").format(series_config))
-		else:
-			# Campo não existe ou está vazio - não é crítico
-			frappe.logger().info("Campo series_configuration não existe ou está vazio")
+				frappe.logger().info(
+					f"✅ Sincronizado com Portugal Series Configuration: {config.name}")
+			else:
+				# Criar configuração se não existir
+				self.create_portugal_series_configuration()
+
+		except Exception as e:
+			frappe.log_error(f"Erro na sincronização: {str(e)}")
+
+	def create_portugal_series_configuration(self):
+		"""
+		✅ NOVO: Criar Portugal Series Configuration correspondente
+		"""
+		try:
+			config = frappe.get_doc({
+				"doctype": "Portugal Series Configuration",
+				"series_name": self.series_name,
+				"company": self.company,
+				"document_type": self.document_type,
+				"prefix": self.prefix,
+				"naming_series": f"{self.prefix}.####",
+				"current_sequence": self.current_number or 1,
+				"is_active": 1,
+				"is_communicated": getattr(self, 'is_communicated', 0),
+				"document_code": self.prefix.split('-')[0] if '-' in self.prefix else self.prefix[
+																					  :2],
+				"year_code": str(getdate().year),
+				"company_code": self.company[:3].upper()
+			})
+
+			config.insert(ignore_permissions=True)
+			self.series_configuration = config.name
+
+			frappe.logger().info(f"✅ Portugal Series Configuration criada: {config.name}")
+
+		except Exception as e:
+			frappe.log_error(f"Erro ao criar configuração: {str(e)}")
+
+	def validate_prefix_format_new(self):
+		"""
+		✅ ADAPTADO: Validar formato compatível com nova abordagem
+		"""
+		if not self.prefix:
+			frappe.throw(_("Series prefix is required"))
+
+		# ✅ ACEITAR AMBOS OS FORMATOS:
+		# Novo: XXYYYY + COMPANY (ex: FT2025DSY)
+		# Antigo: XX-YYYY-COMPANY (ex: FT-2025-DSY)
+
+		new_pattern = r"^[A-Z]{2,4}\d{4}[A-Z0-9]{2,4}$"  # FT2025DSY
+		old_pattern = r"^[A-Z]{2,4}-\d{4}-[A-Z0-9]+$"  # FT-2025-DSY
+
+		if not (re.match(new_pattern, self.prefix) or re.match(old_pattern, self.prefix)):
+			frappe.throw(
+				_("Formato de prefixo inválido. Use: XXYYYY+EMPRESA (ex: FT2025DSY) ou XX-YYYY-EMPRESA (ex: FT-2025-DSY)"))
+
+		# Validar ano no prefixo
+		self.validate_year_in_prefix_new()
+
+	def validate_year_in_prefix_new(self):
+		"""
+		✅ ADAPTADO: Validar ano em ambos os formatos
+		"""
+		if not getattr(self, 'prefix', None):
+			return
+
+		try:
+			# Extrair ano do prefixo
+			if '-' in self.prefix:
+				# Formato antigo: FT-2025-DSY
+				prefix_parts = self.prefix.split('-')
+				if len(prefix_parts) >= 2:
+					year_part = prefix_parts[1]
+			else:
+				# Formato novo: FT2025DSY
+				year_match = re.search(r'\d{4}', self.prefix)
+				if year_match:
+					year_part = year_match.group()
+				else:
+					return
+
+			year = int(year_part)
+			current_year = getdate().year
+
+			if not (current_year - 1 <= year <= current_year + 1):
+				frappe.throw(_("O ano no prefixo deve estar entre {0} e {1}").format(
+					current_year - 1, current_year + 1
+				))
+
+		except (IndexError, ValueError):
+			return
+
+	def validate_prefix_for_document_type_new(self):
+		"""
+		✅ ADAPTADO: Validações alinhadas com nova abordagem
+		"""
+		if not self.document_type or not self.prefix:
+			return
+
+		# ✅ USAR MAPEAMENTO DA NOVA ABORDAGEM
+		from portugal_compliance.regional.portugal import PORTUGAL_DOCUMENT_TYPES
+
+		if self.document_type in PORTUGAL_DOCUMENT_TYPES:
+			valid_code = PORTUGAL_DOCUMENT_TYPES[self.document_type]['code']
+
+			# Extrair código do prefixo
+			if '-' in self.prefix:
+				prefix_code = self.prefix.split('-')[0]
+			else:
+				prefix_code = re.match(r'^[A-Z]{2,4}', self.prefix).group() if re.match(
+					r'^[A-Z]{2,4}', self.prefix) else ""
+
+			if prefix_code != valid_code:
+				frappe.throw(
+					_("Prefixo '{0}' inválido para '{1}'. Use: {2}").format(
+						prefix_code, self.document_type, valid_code
+					))
 
 	def validate_series_name(self):
 		"""Valida nome único da série"""
@@ -66,80 +172,13 @@ class PortugalDocumentSeries(Document):
 		}):
 			frappe.throw(_("Series name '{0}' already exists").format(self.series_name))
 
-	def validate_prefix_format(self):
-		"""Valida formato do prefixo da série"""
-		if not self.prefix:
-			frappe.throw(_("Series prefix is required"))
-
-		# Formato: XX-YYYY-COMPANY ou XXX-YYYY-COMPANY
-		pattern = r"^[A-Z]{2,4}-\d{4}-[A-Z0-9]+$"
-		if not re.match(pattern, self.prefix):
-			frappe.throw(_("Invalid prefix format. Use: XX-YYYY-COMPANY (e.g., FT-2025-COMP)"))
-
-		# Validar ano no prefixo
-		self.validate_year_in_prefix()
-
-	def validate_year_in_prefix(self):
-		"""Valida se o ano no prefixo é válido"""
-		if not getattr(self, 'prefix', None):
-			return
-
-		try:
-			prefix_parts = self.prefix.split('-')
-			if len(prefix_parts) < 2:
-				return
-
-			year_part = prefix_parts[1]
-			year = int(year_part)
-
-			current_year = getdate().year
-
-			if not (current_year - 1 <= year <= current_year + 1):
-				frappe.throw(_("O ano no prefixo deve estar entre {0} e {1}").format(
-					current_year - 1, current_year + 1
-				))
-
-		except (IndexError, ValueError):
-			return
-
-	def validate_prefix_for_document_type(self):
-		"""Valida se o prefixo é adequado para o tipo de documento"""
-		if not self.document_type or not self.prefix:
-			return
-
-		# Mapeamento expandido de prefixos válidos por tipo de documento
-		valid_prefixes = {
-			"Sales Invoice": ["FT", "FS", "FR", "NC", "ND"],
-			"Purchase Invoice": ["FC", "FT"],
-			"Payment Entry": ["RC", "RB"],
-			"Delivery Note": ["GT", "GR"],
-			"Purchase Receipt": ["GR", "GT"],
-			"Journal Entry": ["JE"],
-			"Stock Entry": ["GT"],
-			"Quotation": ["OR", "ORC"],
-			"Sales Order": ["EC", "ECO"],
-			"Purchase Order": ["EF", "EFO"],
-			"Material Request": ["REQ", "MR"]
-		}
-
-		prefix_code = self.prefix.split('-')[0]
-		document_prefixes = valid_prefixes.get(self.document_type, [])
-
-		if document_prefixes and prefix_code not in document_prefixes:
-			frappe.throw(
-				_("Invalid prefix '{0}' for document type '{1}'. Valid prefixes: {2}").format(
-					prefix_code, self.document_type, ", ".join(document_prefixes)
-				))
-
 	def validate_document_type(self):
-		"""Valida tipo de documento - EXPANDIDO para legislação portuguesa"""
-		valid_types = [
-			# Documentos obrigatórios ATCUD
-			"Sales Invoice", "Purchase Invoice", "Payment Entry",
-			"Delivery Note", "Purchase Receipt", "Journal Entry", "Stock Entry",
-			# Documentos adicionais compliance
-			"Quotation", "Sales Order", "Purchase Order", "Material Request"
-		]
+		"""
+		✅ ADAPTADO: Usar tipos da nova abordagem
+		"""
+		from portugal_compliance.regional.portugal import PORTUGAL_DOCUMENT_TYPES
+
+		valid_types = list(PORTUGAL_DOCUMENT_TYPES.keys())
 
 		if self.document_type not in valid_types:
 			frappe.throw(
@@ -213,238 +252,31 @@ class PortugalDocumentSeries(Document):
 		self.last_modified_by_user = frappe.session.user
 
 	def after_insert(self):
-		"""Executado após inserção"""
-		self.create_custom_fields_if_needed()
+		"""
+		✅ ADAPTADO: Não criar custom fields (nova abordagem usa hooks)
+		"""
 		self.log_series_creation()
 		self.update_usage_trend()
 
 	def on_update(self):
 		"""Executado após atualização"""
 		if self.validation_code:
-			self.update_naming_series()
 			self.update_usage_statistics()
+
+		# Sincronizar com Portugal Series Configuration
+		self.sync_with_portugal_series_configuration()
 
 		# Limpar cache se série foi atualizada
 		if self.has_value_changed("validation_code"):
 			frappe.clear_cache()
 
-	def update_naming_series(self):
-		"""Atualiza naming series do doctype"""
-		try:
-			# Obter opções atuais de naming series
-			current_options = frappe.db.get_value("DocType", self.document_type, "autoname") or ""
-			current_list = [opt.strip() for opt in current_options.split('\n') if opt.strip()]
-
-			# Adicionar nova série se não existir
-			new_series = f"{self.prefix}.####"
-			if new_series not in current_list:
-				current_list.append(new_series)
-
-				# Atualizar DocType
-				new_options = '\n'.join(current_list)
-				frappe.db.set_value("DocType", self.document_type, "autoname", new_options)
-
-				# Limpar cache do DocType
-				frappe.clear_cache(doctype=self.document_type)
-
-				frappe.logger().info(
-					f"Naming series updated for {self.document_type}: {new_series}")
-
-		except Exception as e:
-			frappe.log_error(f"Error updating naming series: {str(e)}", "Portugal Document Series")
-
-	def create_custom_fields_if_needed(self):
-		"""✅ SOLUÇÃO DEFINITIVA: Criar campos para TODOS os DocTypes da legislação portuguesa"""
-		try:
-			# ========== CONFIGURAÇÃO COMPLETA DOS DOCTYPES ==========
-			doctypes_config = [
-				# Documentos obrigatórios ATCUD (Decreto-Lei n.º 28/2019)
-				{
-					"doctype": "Sales Invoice",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 1,
-					"atcud_reqd": 1,
-					"priority": "high"
-				},
-				{
-					"doctype": "Purchase Invoice",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 1,
-					"atcud_reqd": 1,
-					"priority": "high"
-				},
-				{
-					"doctype": "Payment Entry",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 1,
-					"atcud_reqd": 1,
-					"priority": "high"
-				},
-				{
-					"doctype": "Delivery Note",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 1,
-					"atcud_reqd": 1,
-					"priority": "high"
-				},
-				{
-					"doctype": "Purchase Receipt",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 1,
-					"atcud_reqd": 1,
-					"priority": "high"
-				},
-				{
-					"doctype": "Journal Entry",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 0,
-					"atcud_reqd": 1,
-					"priority": "medium"
-				},
-				{
-					"doctype": "Stock Entry",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 0,
-					"atcud_reqd": 1,
-					"priority": "medium"
-				},
-				# Documentos adicionais compliance
-				{
-					"doctype": "Quotation",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 0,
-					"atcud_reqd": 0,
-					"priority": "low"
-				},
-				{
-					"doctype": "Sales Order",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 0,
-					"atcud_reqd": 0,
-					"priority": "low"
-				},
-				{
-					"doctype": "Purchase Order",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 0,
-					"atcud_reqd": 0,
-					"priority": "low"
-				},
-				{
-					"doctype": "Material Request",
-					"insert_after": "naming_series",
-					"portugal_series_reqd": 0,
-					"atcud_reqd": 0,
-					"priority": "low"
-				}
-			]
-
-			# ========== CRIAR CAMPOS PARA TODOS OS DOCTYPES ==========
-			for config in doctypes_config:
-				try:
-					doctype = config["doctype"]
-
-					# ✅ 1. CAMPO PORTUGAL_SERIES
-					series_field_name = f"{doctype}-portugal_series"
-					if not frappe.db.exists("Custom Field", series_field_name):
-						portugal_series_field = frappe.get_doc({
-							"doctype": "Custom Field",
-							"dt": doctype,
-							"fieldname": "portugal_series",
-							"label": "Portugal Series",
-							"fieldtype": "Link",
-							"options": "Portugal Document Series",
-							"insert_after": config["insert_after"],
-							"reqd": config["portugal_series_reqd"],
-							"in_list_view": 1 if config["portugal_series_reqd"] else 0,
-							"bold": 1 if config["portugal_series_reqd"] else 0,
-							"description": f"Série portuguesa para compliance fiscal - {config['priority']} priority"
-						})
-						portugal_series_field.insert(ignore_permissions=True)
-						frappe.logger().info(f"✅ Campo portugal_series criado para {doctype}")
-
-					# ✅ 2. CAMPO ATCUD_CODE
-					atcud_field_name = f"{doctype}-atcud_code"
-					if not frappe.db.exists("Custom Field", atcud_field_name):
-						atcud_field = frappe.get_doc({
-							"doctype": "Custom Field",
-							"dt": doctype,
-							"fieldname": "atcud_code",
-							"label": "ATCUD Code",
-							"fieldtype": "Data",
-							"insert_after": "portugal_series",
-							"read_only": 1,
-							"print_hide": 0,
-							"bold": 1 if config["atcud_reqd"] else 0,
-							"in_list_view": 1 if config["atcud_reqd"] else 0,
-							"description": "Código Único de Documento - obrigatório em Portugal"
-						})
-						atcud_field.insert(ignore_permissions=True)
-						frappe.logger().info(f"✅ Campo atcud_code criado para {doctype}")
-
-					# ✅ 3. CAMPO ATCUD_DISPLAY (para impressão)
-					atcud_display_field_name = f"{doctype}-atcud_display"
-					if not frappe.db.exists("Custom Field", atcud_display_field_name):
-						atcud_display_field = frappe.get_doc({
-							"doctype": "Custom Field",
-							"dt": doctype,
-							"fieldname": "atcud_display",
-							"label": "ATCUD Display",
-							"fieldtype": "Data",
-							"insert_after": "atcud_code",
-							"read_only": 1,
-							"print_hide": 0,
-							"hidden": 1,  # Oculto na interface, visível na impressão
-							"description": "ATCUD formatado para exibição (ATCUD:CODIGO-NUMERO)"
-						})
-						atcud_display_field.insert(ignore_permissions=True)
-						frappe.logger().info(f"✅ Campo atcud_display criado para {doctype}")
-
-					# ✅ 4. CAMPO PORTUGAL_COMPLIANCE_STATUS
-					compliance_field_name = f"{doctype}-portugal_compliance_status"
-					if not frappe.db.exists("Custom Field", compliance_field_name):
-						compliance_field = frappe.get_doc({
-							"doctype": "Custom Field",
-							"dt": doctype,
-							"fieldname": "portugal_compliance_status",
-							"label": "Portugal Compliance Status",
-							"fieldtype": "Select",
-							"options": "Pending\nCompliant\nNon-Compliant\nExempt",
-							"insert_after": "atcud_display",
-							"default": "Pending",
-							"read_only": 1,
-							"hidden": 1,
-							"description": "Status de compliance português para este documento"
-						})
-						compliance_field.insert(ignore_permissions=True)
-						frappe.logger().info(
-							f"✅ Campo portugal_compliance_status criado para {doctype}")
-
-					frappe.logger().info(f"🎯 Todos os campos criados com sucesso para {doctype}")
-
-				except Exception as e:
-					frappe.log_error(f"❌ Erro ao criar campos para {config['doctype']}: {str(e)}",
-									 "Portugal Document Series - Custom Fields")
-					# Continuar com próximo DocType mesmo se um falhar
-					continue
-
-			# ========== COMMIT E LIMPEZA DE CACHE ==========
-			frappe.db.commit()
-			frappe.clear_cache()
-
-			frappe.logger().info("🎉 TODOS os campos customizados criados com sucesso!")
-			frappe.logger().info(
-				"📋 DocTypes atualizados: Sales Invoice, Purchase Invoice, Payment Entry, Delivery Note, Purchase Receipt, Journal Entry, Stock Entry, Quotation, Sales Order, Purchase Order, Material Request")
-
-		except Exception as e:
-			frappe.log_error(f"❌ Erro geral ao criar campos customizados: {str(e)}",
-							 "Portugal Document Series - Custom Fields")
-
-	# ========== MÉTODOS DE GERAÇÃO DE ATCUD PARA DOCUMENTOS ==========
+	# ========== MÉTODOS DE GERAÇÃO DE ATCUD ADAPTADOS ==========
 
 	@frappe.whitelist()
 	def generate_atcud_for_document(self, document_sequence=None):
-		"""Gerar ATCUD completo para documento específico - CORRIGIDO"""
+		"""
+		✅ ADAPTADO: Gerar ATCUD usando nova abordagem
+		"""
 		try:
 			if not self.validation_code:
 				frappe.throw(
@@ -454,23 +286,15 @@ class PortugalDocumentSeries(Document):
 			if document_sequence is None:
 				document_sequence = self.get_next_number()
 
-			# ✅ FORMATO OFICIAL CORRETO: CODIGO_VALIDACAO-NUMERO_SEQUENCIAL
-			# Baseado nos search results: "TES123TE-4561"
-			validation_code = str(self.validation_code).strip()
-			sequence_formatted = f"{document_sequence:08d}"  # 8 dígitos com zeros à esquerda
-
-			atcud_code = f"{validation_code}-{sequence_formatted}"
+			# ✅ FORMATO NOVO: 0.SEQUENCIAL (conforme Portaria 195/2020)
+			atcud_code = f"0.{document_sequence}"
 			atcud_display = f"ATCUD:{atcud_code}"
-
-			# Validar formato gerado
-			if not self.validate_atcud_format(atcud_code):
-				frappe.throw(_("Erro na geração do ATCUD: formato inválido"))
 
 			return {
 				"success": True,
-				"atcud_code": atcud_code,  # Ex: AAJFJ6VHXK-00000001
-				"atcud_display": atcud_display,  # Ex: ATCUD:AAJFJ6VHXK-00000001
-				"validation_code": validation_code,  # Ex: AAJFJ6VHXK
+				"atcud_code": atcud_code,  # Ex: 0.1
+				"atcud_display": atcud_display,  # Ex: ATCUD:0.1
+				"validation_code": self.validation_code,
 				"sequence_number": document_sequence,
 				"series_name": self.series_name
 			}
@@ -478,33 +302,6 @@ class PortugalDocumentSeries(Document):
 		except Exception as e:
 			frappe.log_error(f"Error generating ATCUD: {str(e)}")
 			return {"success": False, "error": str(e)}
-
-	def validate_atcud_format(self, atcud_code):
-		"""Validar formato do ATCUD gerado"""
-		try:
-			# Baseado nos search results: deve ter hífen separando código e sequência
-			if "-" not in atcud_code:
-				return False
-
-			parts = atcud_code.split("-")
-			if len(parts) != 2:
-				return False
-
-			validation_code, sequence = parts
-
-			# Código de validação: 8+ caracteres, sem 0 e 1, sem acentos
-			if len(validation_code) < 8:
-				return False
-
-			# Sequência: deve ser numérica
-			try:
-				int(sequence)
-				return True
-			except ValueError:
-				return False
-
-		except Exception:
-			return False
 
 	def get_next_number(self):
 		"""Obtém próximo número da sequência de forma thread-safe"""
@@ -525,43 +322,29 @@ class PortugalDocumentSeries(Document):
 			frappe.db.sql("SELECT RELEASE_LOCK(%s)", (f"series_lock_{self.name}",))
 			return self.current_number
 
-	# ========== MÉTODOS DE ESTATÍSTICAS (mantidos do código original) ==========
+	# ========== MÉTODOS DE ESTATÍSTICAS (SIMPLIFICADOS) ==========
 
 	def update_usage_statistics(self):
 		"""Atualiza estatísticas de uso da série"""
 		try:
-			# Contar documentos que usam esta série
+			# Contar documentos que usam esta série via naming_series
+			naming_series = f"{self.prefix}.####"
+
 			total_docs = frappe.db.count(self.document_type, {
-				"portugal_series": self.name,
-				"docstatus": ["!=", 2]  # Excluir cancelados
+				"naming_series": naming_series,
+				"company": self.company,
+				"docstatus": ["!=", 2]
 			})
 
-			# Último documento criado
-			last_doc_data = frappe.db.get_value(self.document_type, {
-				"portugal_series": self.name,
-				"docstatus": ["!=", 2]
-			}, ["creation", "posting_date"], order_by="creation desc")
+			# Atualizar usando db_set para evitar loops
+			self.db_set("total_documents_issued", total_docs, update_modified=False)
 
 			# Calcular uso médio mensal
 			monthly_usage = self.calculate_monthly_usage()
-
-			# Calcular projeção anual
-			annual_projection = self.calculate_annual_projection(monthly_usage)
-
-			# Atualizar campos usando db_set para evitar loops
-			self.db_set("total_documents_issued", total_docs, update_modified=False)
-
-			if last_doc_data:
-				last_date = last_doc_data[1] if last_doc_data[1] else last_doc_data[0].date()
-				self.db_set("last_document_date", last_date, update_modified=False)
-
 			if monthly_usage:
 				self.db_set("average_monthly_usage", monthly_usage, update_modified=False)
 
-			if annual_projection:
-				self.db_set("projected_annual_usage", annual_projection, update_modified=False)
-
-			# Atualizar tendência de uso
+			# Atualizar tendência
 			self.update_usage_trend()
 
 		except Exception as e:
@@ -574,6 +357,7 @@ class PortugalDocumentSeries(Document):
 
 			end_date = today()
 			start_date = end_date - relativedelta(months=12)
+			naming_series = f"{self.prefix}.####"
 
 			monthly_counts = frappe.db.sql("""
                 SELECT
@@ -581,11 +365,13 @@ class PortugalDocumentSeries(Document):
                     MONTH(posting_date) as month,
                     COUNT(*) as count
                 FROM `tab{doctype}`
-                WHERE portugal_series = %s
+                WHERE naming_series = %s
+                AND company = %s
                 AND posting_date BETWEEN %s AND %s
                 AND docstatus != 2
                 GROUP BY YEAR(posting_date), MONTH(posting_date)
-            """.format(doctype=self.document_type), (self.name, start_date, end_date),
+            """.format(doctype=self.document_type),
+										   (naming_series, self.company, start_date, end_date),
 										   as_dict=True)
 
 			if monthly_counts:
@@ -599,15 +385,6 @@ class PortugalDocumentSeries(Document):
 			frappe.log_error(f"Error calculating monthly usage: {str(e)}")
 			return 0
 
-	def calculate_annual_projection(self, monthly_usage):
-		"""Calcula projeção anual baseada no uso médio mensal"""
-		try:
-			if monthly_usage:
-				return cint(monthly_usage * 12)
-			return 0
-		except:
-			return 0
-
 	def update_usage_trend(self):
 		"""Atualiza tendência de uso"""
 		try:
@@ -615,17 +392,6 @@ class PortugalDocumentSeries(Document):
 				trend = "New"
 			elif self.total_documents_issued < 10:
 				trend = "New"
-			elif self.average_monthly_usage:
-				# Comparar últimos 3 meses com 3 meses anteriores
-				recent_usage = self.get_recent_usage(3)
-				previous_usage = self.get_previous_usage(3, 6)
-
-				if recent_usage > previous_usage * 1.1:
-					trend = "Increasing"
-				elif recent_usage < previous_usage * 0.9:
-					trend = "Decreasing"
-				else:
-					trend = "Stable"
 			else:
 				trend = "Stable"
 
@@ -633,44 +399,6 @@ class PortugalDocumentSeries(Document):
 
 		except Exception as e:
 			frappe.log_error(f"Error updating usage trend: {str(e)}")
-
-	def get_recent_usage(self, months):
-		"""Obtém uso dos últimos N meses"""
-		try:
-			from dateutil.relativedelta import relativedelta
-
-			end_date = today()
-			start_date = end_date - relativedelta(months=months)
-
-			count = frappe.db.count(self.document_type, {
-				"portugal_series": self.name,
-				"posting_date": ["between", [start_date, end_date]],
-				"docstatus": ["!=", 2]
-			})
-
-			return count
-		except:
-			return 0
-
-	def get_previous_usage(self, months, offset):
-		"""Obtém uso de período anterior"""
-		try:
-			from dateutil.relativedelta import relativedelta
-
-			end_date = today() - relativedelta(months=offset)
-			start_date = end_date - relativedelta(months=months)
-
-			count = frappe.db.count(self.document_type, {
-				"portugal_series": self.name,
-				"posting_date": ["between", [start_date, end_date]],
-				"docstatus": ["!=", 2]
-			})
-
-			return count
-		except:
-			return 0
-
-	# ========== MÉTODOS AUXILIARES ==========
 
 	def log_series_creation(self):
 		"""Registra log da criação da série"""
@@ -690,54 +418,17 @@ class PortugalDocumentSeries(Document):
 		self.current_number = cint(new_start)
 		self.save()
 
-		# Log da ação
-		frappe.get_doc({
-			"doctype": "Comment",
-			"comment_type": "Info",
-			"reference_doctype": self.doctype,
-			"reference_name": self.name,
-			"content": f"Sequence reset from {old_number} to {new_start} by {frappe.session.user}"
-		}).insert(ignore_permissions=True)
+		# Sincronizar com Portugal Series Configuration
+		if hasattr(self, 'series_configuration') and self.series_configuration:
+			try:
+				config = frappe.get_doc("Portugal Series Configuration", self.series_configuration)
+				config.current_sequence = cint(new_start)
+				config.save(ignore_permissions=True)
+			except:
+				pass
 
 		frappe.msgprint(
 			_("Sequence reset successfully from {0} to {1}").format(old_number, new_start))
-
-	@frappe.whitelist()
-	def get_usage_statistics(self):
-		"""Retorna estatísticas detalhadas de uso"""
-		try:
-			stats = {
-				"series_name": self.series_name,
-				"prefix": self.prefix,
-				"document_type": self.document_type,
-				"current_number": self.current_number,
-				"validation_code": self.validation_code,
-				"total_documents": self.total_documents_issued or 0,
-				"last_document_date": self.last_document_date,
-				"average_monthly_usage": self.average_monthly_usage or 0,
-				"projected_annual_usage": self.projected_annual_usage or 0,
-				"usage_trend": self.usage_trend or "New",
-				"available_numbers": 99999999 - (self.current_number or 1)
-			}
-
-			# Obter último documento
-			if self.total_documents_issued and self.total_documents_issued > 0:
-				last_doc = frappe.db.get_value(self.document_type, {
-					"portugal_series": self.name,
-					"docstatus": ["!=", 2]
-				}, ["name", "creation"], order_by="creation desc")
-
-				if last_doc:
-					stats["last_document"] = {
-						"name": last_doc[0],
-						"creation": last_doc[1]
-					}
-
-			return stats
-
-		except Exception as e:
-			frappe.log_error(f"Error getting usage statistics: {str(e)}")
-			return {}
 
 	def on_trash(self):
 		"""Executado antes de eliminar"""
@@ -745,66 +436,31 @@ class PortugalDocumentSeries(Document):
 		if self.is_series_in_use():
 			frappe.throw(_("Cannot delete series that is being used by documents"))
 
-		# Log da eliminação
+		# Eliminar Portugal Series Configuration correspondente
+		if hasattr(self, 'series_configuration') and self.series_configuration:
+			try:
+				frappe.delete_doc("Portugal Series Configuration", self.series_configuration,
+								  ignore_permissions=True)
+			except:
+				pass
+
 		frappe.logger().info(f"Portugal Document Series deleted: {self.series_name}")
 
 	def is_series_in_use(self):
 		"""Verifica se a série está sendo usada por documentos"""
 		try:
+			naming_series = f"{self.prefix}.####"
 			count = frappe.db.count(self.document_type, {
-				"portugal_series": self.name,
+				"naming_series": naming_series,
+				"company": self.company,
 				"docstatus": ["!=", 2]
 			})
 			return count > 0
 		except:
 			return False
 
-	@frappe.whitelist()
-	def get_usage_stats(self):
-		"""Método whitelisted para compatibilidade com JS"""
-		return self.get_usage_statistics()
 
-	# ========== MÉTODOS DE COMPLIANCE AUTOMÁTICO ==========
-
-	@frappe.whitelist()
-	def validate_all_documents_compliance(self):
-		"""Validar compliance de todos os documentos desta série"""
-		try:
-			# Buscar todos os documentos desta série
-			documents = frappe.get_all(self.document_type,
-									   filters={"portugal_series": self.name,
-												"docstatus": ["!=", 2]},
-									   fields=["name", "atcud_code", "portugal_compliance_status"])
-
-			compliance_stats = {
-				"total": len(documents),
-				"compliant": 0,
-				"non_compliant": 0,
-				"pending": 0
-			}
-
-			for doc in documents:
-				if doc.get("atcud_code") and doc.get("portugal_compliance_status") == "Compliant":
-					compliance_stats["compliant"] += 1
-				elif not doc.get("atcud_code"):
-					compliance_stats["non_compliant"] += 1
-				else:
-					compliance_stats["pending"] += 1
-
-			return {
-				"success": True,
-				"compliance_stats": compliance_stats,
-				"compliance_rate": (
-						compliance_stats["compliant"] / compliance_stats["total"] * 100) if
-				compliance_stats["total"] > 0 else 0
-			}
-
-		except Exception as e:
-			frappe.log_error(f"Error validating documents compliance: {str(e)}")
-			return {"success": False, "error": str(e)}
-
-
-# ========== FUNÇÕES GLOBAIS PARA INTEGRAÇÃO COM DOCUMENTOS ==========
+# ========== FUNÇÕES GLOBAIS ADAPTADAS ==========
 
 @frappe.whitelist()
 def get_atcud_for_document(series_name, document_sequence=None):
@@ -835,40 +491,27 @@ def get_series_for_document_type(document_type, company=None):
 
 
 @frappe.whitelist()
-def create_all_custom_fields():
-	"""API para criar campos customizados em todos os DocTypes (uso manual)"""
+def sync_all_series_with_configuration():
+	"""
+	✅ NOVO: Sincronizar todas as séries com Portugal Series Configuration
+	"""
 	try:
-		# Criar uma série temporária apenas para executar o método
-		temp_series = frappe.new_doc("Portugal Document Series")
-		temp_series.create_custom_fields_if_needed()
+		series_list = frappe.get_all("Portugal Document Series",
+									 fields=["name"])
 
-		return {"success": True, "message": "Todos os campos customizados foram criados"}
-	except Exception as e:
-		return {"success": False, "error": str(e)}
-
-
-@frappe.whitelist()
-def validate_portugal_compliance_setup():
-	"""Validar se setup de compliance português está correto"""
-	try:
-		# Verificar se todos os campos existem
-		required_doctypes = [
-			"Sales Invoice", "Purchase Invoice", "Payment Entry",
-			"Delivery Note", "Purchase Receipt", "Journal Entry", "Stock Entry"
-		]
-
-		missing_fields = []
-		for doctype in required_doctypes:
-			if not frappe.db.has_column(doctype, "portugal_series"):
-				missing_fields.append(f"{doctype}.portugal_series")
-			if not frappe.db.has_column(doctype, "atcud_code"):
-				missing_fields.append(f"{doctype}.atcud_code")
+		synced_count = 0
+		for series in series_list:
+			try:
+				series_doc = frappe.get_doc("Portugal Document Series", series.name)
+				series_doc.sync_with_portugal_series_configuration()
+				synced_count += 1
+			except Exception as e:
+				frappe.log_error(f"Erro ao sincronizar série {series.name}: {str(e)}")
 
 		return {
-			"success": len(missing_fields) == 0,
-			"missing_fields": missing_fields,
-			"message": "Setup completo" if len(
-				missing_fields) == 0 else f"Faltam {len(missing_fields)} campos"
+			"success": True,
+			"message": f"{synced_count} séries sincronizadas com sucesso",
+			"synced_count": synced_count
 		}
 
 	except Exception as e:
