@@ -4,7 +4,7 @@
 
 """
 Tarefas que executam diariamente - VERSÃO ATUALIZADA
-Integração completa com Portugal Compliance
+Integração completa com Portugal Compliance - NOVA ABORDAGEM
 """
 
 import frappe
@@ -16,17 +16,18 @@ import json
 
 def execute():
 	"""
-	Função principal executada diariamente - ATUALIZADA
+	Função principal executada diariamente - ATUALIZADA PARA NOVA ABORDAGEM
 	"""
 	try:
-		frappe.logger().info("🇵🇹 Portugal Compliance: Executing daily scheduled tasks")
+		frappe.logger().info(
+			"🇵🇹 Portugal Compliance: Executing daily scheduled tasks - NEW APPROACH")
 
 		# Verificar se o módulo está ativo
 		if not is_portugal_compliance_enabled():
 			frappe.logger().info("Portugal Compliance not enabled, skipping daily tasks")
 			return
 
-		# ✅ TAREFAS PRINCIPAIS (MANTIDAS + MELHORADAS)
+		# ✅ TAREFAS PRINCIPAIS (ADAPTADAS PARA NOVA ABORDAGEM)
 		generate_daily_report()
 		check_series_expiration()
 		validate_daily_sequences()
@@ -36,17 +37,19 @@ def execute():
 		backup_critical_data()
 		send_daily_notifications()
 
-		# ✅ NOVAS TAREFAS ADICIONADAS
-		validate_prefix_formats()
+		# ✅ NOVAS TAREFAS ADAPTADAS PARA NAMING SERIES NATIVA
+		validate_naming_series_formats()
 		check_atcud_sequence_integrity()
 		monitor_communication_failures()
 		update_series_trends()
 		check_naming_series_consistency()
-		validate_custom_fields_integrity()
+		validate_essential_custom_fields_integrity()
 		cleanup_failed_communications()
 		generate_compliance_metrics()
+		sync_portugal_series_configurations()
 
-		frappe.logger().info("🇵🇹 Portugal Compliance: Daily tasks completed successfully")
+		frappe.logger().info(
+			"🇵🇹 Portugal Compliance: Daily tasks completed successfully - NEW APPROACH")
 
 	except Exception as e:
 		frappe.log_error(f"Error in portugal_compliance.tasks.daily: {str(e)}")
@@ -54,7 +57,7 @@ def execute():
 
 def is_portugal_compliance_enabled():
 	"""
-	Verifica se o módulo Portugal Compliance está ativo - MELHORADO
+	Verifica se o módulo Portugal Compliance está ativo - ADAPTADO
 	"""
 	try:
 		# Verificar se há empresas portuguesas com compliance ativado
@@ -63,7 +66,7 @@ def is_portugal_compliance_enabled():
 			"portugal_compliance_enabled": 1
 		})
 
-		# Verificar se há séries ativas
+		# ✅ ADAPTADO: Verificar Portugal Series Configuration em vez de Document Series
 		active_series = frappe.db.count("Portugal Series Configuration", {
 			"is_active": 1
 		})
@@ -76,12 +79,12 @@ def is_portugal_compliance_enabled():
 
 def generate_daily_report():
 	"""
-	Gera relatório diário de atividades - EXPANDIDO
+	Gera relatório diário de atividades - ADAPTADO PARA NOVA ABORDAGEM
 	"""
 	try:
 		today_date = today()
 
-		# ✅ ESTATÍSTICAS EXPANDIDAS
+		# ✅ ESTATÍSTICAS ADAPTADAS PARA NAMING SERIES NATIVA
 		daily_stats = {
 			"date": today_date,
 			"atcud_generated": frappe.db.count("ATCUD Log", {
@@ -99,11 +102,11 @@ def generate_daily_report():
 			"companies_active": frappe.db.count("Company", {
 				"portugal_compliance_enabled": 1
 			}),
-			# ✅ NOVAS MÉTRICAS
+			# ✅ MÉTRICAS ADAPTADAS PARA NOVA ABORDAGEM
 			"series_with_new_format": count_series_with_new_format(),
 			"communication_success_rate": calculate_communication_success_rate(),
 			"avg_atcud_generation_time": calculate_avg_atcud_time(),
-			"prefix_format_compliance": calculate_prefix_compliance(),
+			"naming_series_compliance": calculate_naming_series_compliance(),
 			"naming_series_consistency": check_naming_series_health()
 		}
 
@@ -112,20 +115,20 @@ def generate_daily_report():
 
 		# Log se houver atividade significativa
 		if daily_stats["atcud_generated"] > 0 or daily_stats["series_communicated"] > 0:
-			frappe.logger().info(f"📊 Daily Report: {daily_stats}")
+			frappe.logger().info(f"📊 Daily Report (New Approach): {daily_stats}")
 
 	except Exception as e:
 		frappe.log_error(f"Error generating daily report: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Contar séries com novo formato
+# ✅ FUNÇÃO ADAPTADA: Contar séries com formato correto
 def count_series_with_new_format():
 	"""
-	Conta séries que usam o novo formato XX-YYYY-COMPANY
+	Conta séries que usam o formato correto XXYYYY+EMPRESA
 	"""
 	try:
-		# Padrão do novo formato
-		new_format_pattern = r'^[A-Z]{2,4}-\d{4}-[A-Z0-9]{2,4}$'
+		# ✅ ADAPTADO: Padrão do formato correto (sem hífens)
+		new_format_pattern = r'^[A-Z]{2,4}\d{4}[A-Z0-9]{2,4}$'
 
 		all_series = frappe.db.get_all("Portugal Series Configuration",
 									   fields=["name", "prefix"])
@@ -144,20 +147,47 @@ def count_series_with_new_format():
 		return 0
 
 
-# ✅ NOVA FUNÇÃO: Calcular taxa de sucesso de comunicação
+# ✅ FUNÇÃO ADAPTADA: Calcular compliance de naming series
+def calculate_naming_series_compliance():
+	"""
+	Calcula percentagem de séries usando naming series nativa corretamente
+	"""
+	try:
+		total_series = frappe.db.count("Portugal Series Configuration")
+
+		# Contar séries que têm naming_series definida corretamente
+		correct_naming_series = 0
+
+		all_series = frappe.db.get_all("Portugal Series Configuration",
+									   fields=["prefix", "naming_series"])
+
+		for series in all_series:
+			if series.prefix and series.naming_series:
+				expected_naming = f"{series.prefix}.####"
+				if series.naming_series == expected_naming:
+					correct_naming_series += 1
+
+		if total_series > 0:
+			return round((correct_naming_series / total_series) * 100, 2)
+		return 100
+
+	except Exception:
+		return 0
+
+
+# ✅ FUNÇÃO ADAPTADA: Calcular taxa de sucesso de comunicação
 def calculate_communication_success_rate():
 	"""
-	Calcula taxa de sucesso das comunicações AT
+	Calcula taxa de sucesso das comunicações AT - ADAPTADA
 	"""
 	try:
 		total_attempts = frappe.db.count("Portugal Series Configuration", {
-			"enable_at_communication": 1
+			"is_active": 1
 		})
 
 		successful_communications = frappe.db.count("Portugal Series Configuration", {
-			"enable_at_communication": 1,
-			"is_communicated": 1,
-			"communication_status": "Success"
+			"is_active": 1,
+			"is_communicated": 1
 		})
 
 		if total_attempts > 0:
@@ -168,13 +198,11 @@ def calculate_communication_success_rate():
 		return 0
 
 
-# ✅ NOVA FUNÇÃO: Calcular tempo médio de geração ATCUD
 def calculate_avg_atcud_time():
 	"""
-	Calcula tempo médio de geração de ATCUD
+	Calcula tempo médio de geração de ATCUD - MANTIDA
 	"""
 	try:
-		# Esta é uma métrica estimada baseada nos logs
 		recent_logs = frappe.db.get_all("ATCUD Log",
 										filters={
 											"creation": [">=", add_days(today(), -7)],
@@ -183,35 +211,17 @@ def calculate_avg_atcud_time():
 										fields=["creation", "generation_date"])
 
 		if len(recent_logs) > 0:
-			# Estimativa baseada na quantidade de logs recentes
-			return round(len(recent_logs) / 7, 2)  # Média por dia
+			return round(len(recent_logs) / 7, 2)
 		return 0
 
 	except Exception:
 		return 0
 
 
-# ✅ NOVA FUNÇÃO: Calcular compliance de formato de prefixo
-def calculate_prefix_compliance():
-	"""
-	Calcula percentagem de séries com formato correto
-	"""
-	try:
-		total_series = frappe.db.count("Portugal Series Configuration")
-		new_format_series = count_series_with_new_format()
-
-		if total_series > 0:
-			return round((new_format_series / total_series) * 100, 2)
-		return 100
-
-	except Exception:
-		return 0
-
-
-# ✅ NOVA FUNÇÃO: Verificar saúde das naming series
+# ✅ FUNÇÃO ADAPTADA: Verificar saúde das naming series
 def check_naming_series_health():
 	"""
-	Verifica consistência das naming series
+	Verifica consistência das naming series com nova abordagem
 	"""
 	try:
 		series_with_issues = 0
@@ -219,22 +229,16 @@ def check_naming_series_health():
 
 		active_series = frappe.db.get_all("Portugal Series Configuration",
 										  filters={"is_active": 1},
-										  fields=["name", "prefix", "document_type"])
+										  fields=["name", "prefix", "document_type",
+												  "naming_series"])
 
 		for series in active_series:
 			total_series += 1
 
-			# Verificar se naming series está configurada corretamente
+			# ✅ ADAPTADO: Verificar se naming series está configurada corretamente
 			expected_naming = f"{series.prefix}.####"
 
-			try:
-				doctype_meta = frappe.get_meta(series.document_type)
-				autoname_options = getattr(doctype_meta, 'autoname', '') or ''
-
-				if expected_naming not in autoname_options:
-					series_with_issues += 1
-
-			except Exception:
+			if not series.naming_series or series.naming_series != expected_naming:
 				series_with_issues += 1
 
 		if total_series > 0:
@@ -249,15 +253,15 @@ def check_naming_series_health():
 
 def get_daily_document_count():
 	"""
-	Conta documentos processados hoje - MELHORADO
+	Conta documentos processados hoje - ADAPTADO PARA NAMING SERIES
 	"""
 	try:
 		today_date = today()
 		total_docs = 0
 
-		# ✅ TIPOS DE DOCUMENTOS EXPANDIDOS
+		# ✅ TIPOS DE DOCUMENTOS MANTIDOS
 		doctypes = [
-			"Sales Invoice", "Purchase Invoice", "Payment Entry",
+			"Sales Invoice", "POS Invoice", "Purchase Invoice", "Payment Entry",
 			"Delivery Note", "Purchase Receipt", "Journal Entry",
 			"Stock Entry", "Quotation", "Sales Order", "Purchase Order",
 			"Material Request"
@@ -265,6 +269,7 @@ def get_daily_document_count():
 
 		for doctype in doctypes:
 			if frappe.db.table_exists(f"tab{doctype}"):
+				# ✅ ADAPTADO: Contar documentos com ATCUD (não portugal_series)
 				count = frappe.db.count(doctype, {
 					"creation": [">=", today_date],
 					"atcud_code": ["is", "set"]
@@ -280,10 +285,9 @@ def get_daily_document_count():
 
 def store_daily_report(stats):
 	"""
-	Armazena relatório diário na base de dados - MELHORADO
+	Armazena relatório diário na base de dados - ADAPTADO
 	"""
 	try:
-		# ✅ USAR CACHE SE DOCTYPE NÃO EXISTIR
 		cache_key = f"portugal_compliance_daily_report_{stats['date']}"
 
 		# Tentar armazenar em DocType se existir
@@ -302,7 +306,7 @@ def store_daily_report(stats):
 					"companies_active": stats["companies_active"],
 					"series_with_new_format": stats.get("series_with_new_format", 0),
 					"communication_success_rate": stats.get("communication_success_rate", 0),
-					"prefix_format_compliance": stats.get("prefix_format_compliance", 0),
+					"naming_series_compliance": stats.get("naming_series_compliance", 0),
 					"last_updated": now()
 				})
 			else:
@@ -322,51 +326,47 @@ def store_daily_report(stats):
 		frappe.log_error(f"Error storing daily report: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Validar formatos de prefixo
-def validate_prefix_formats():
+# ✅ FUNÇÃO ADAPTADA: Validar formatos de naming series
+def validate_naming_series_formats():
 	"""
-	Valida e reporta formatos de prefixo incorretos
+	Valida e reporta formatos de naming series incorretos
 	"""
 	try:
 		import re
-		new_format_pattern = r'^[A-Z]{2,4}-\d{4}-[A-Z0-9]{2,4}$'
-		old_format_pattern = r'^[A-Z]{2,4}-\d{4}-[A-Z0-9]+$'
+		correct_pattern = r'^[A-Z]{2,4}\d{4}[A-Z0-9]{2,4}\.####$'
 
 		all_series = frappe.db.get_all("Portugal Series Configuration",
-									   fields=["name", "prefix", "company"])
+									   fields=["name", "prefix", "naming_series", "company"])
 
-		invalid_prefixes = []
-		legacy_prefixes = []
+		invalid_naming_series = []
+		missing_naming_series = []
 
 		for series in all_series:
-			if series.prefix:
-				if re.match(new_format_pattern, series.prefix):
-					continue  # Formato correto
-				elif re.match(old_format_pattern, series.prefix):
-					legacy_prefixes.append(series)
-				else:
-					invalid_prefixes.append(series)
+			if not series.naming_series:
+				missing_naming_series.append(series)
+			elif not re.match(correct_pattern, series.naming_series):
+				invalid_naming_series.append(series)
 
 		# Log resultados
-		if invalid_prefixes:
+		if invalid_naming_series:
 			frappe.logger().warning(
-				f"⚠️ Found {len(invalid_prefixes)} series with invalid prefix format")
+				f"⚠️ Found {len(invalid_naming_series)} series with invalid naming series format")
 
-		if legacy_prefixes:
-			frappe.logger().info(f"📋 Found {len(legacy_prefixes)} series with legacy format")
+		if missing_naming_series:
+			frappe.logger().warning(
+				f"⚠️ Found {len(missing_naming_series)} series without naming series")
 
 		# Criar notificações se necessário
-		if len(invalid_prefixes) > 5:  # Muitos prefixos inválidos
-			create_prefix_format_alert(invalid_prefixes, legacy_prefixes)
+		if len(invalid_naming_series) > 5 or len(missing_naming_series) > 5:
+			create_naming_series_format_alert(invalid_naming_series, missing_naming_series)
 
 	except Exception as e:
-		frappe.log_error(f"Error validating prefix formats: {str(e)}")
+		frappe.log_error(f"Error validating naming series formats: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Verificar integridade de sequências ATCUD
 def check_atcud_sequence_integrity():
 	"""
-	Verifica integridade das sequências ATCUD
+	Verifica integridade das sequências ATCUD - MANTIDA
 	"""
 	try:
 		# Verificar duplicações de ATCUD
@@ -381,16 +381,14 @@ def check_atcud_sequence_integrity():
 
 		if duplicate_atcuds:
 			frappe.logger().error(f"❌ Found {len(duplicate_atcuds)} duplicate ATCUD codes")
-
-			# Criar alerta crítico
 			create_atcud_integrity_alert(duplicate_atcuds)
 
-		# Verificar formato de ATCUDs
+		# ✅ ADAPTADO: Verificar formato de ATCUDs (novo formato: 0.SEQUENCIAL)
 		invalid_atcuds = frappe.db.sql("""
 									   SELECT name, atcud_code, document_type
 									   FROM `tabATCUD Log`
 									   WHERE atcud_code IS NOT NULL
-										 AND atcud_code NOT REGEXP '^[A-Z0-9]+-[0-9]{8}$'
+										 AND atcud_code NOT REGEXP '^0\\.[0-9]+$'
 			  AND creation >= %s
 									   """, add_days(today(), -7), as_dict=True)
 
@@ -401,35 +399,28 @@ def check_atcud_sequence_integrity():
 		frappe.log_error(f"Error checking ATCUD sequence integrity: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Monitorizar falhas de comunicação
 def monitor_communication_failures():
 	"""
-	Monitoriza e reporta falhas de comunicação AT
+	Monitoriza e reporta falhas de comunicação AT - ADAPTADA
 	"""
 	try:
 		# Séries com falhas recentes
 		failed_communications = frappe.db.get_all("Portugal Series Configuration",
 												  filters={
-													  "communication_status": "Failed",
-													  "enable_at_communication": 1,
+													  "is_active": 1,
+													  "is_communicated": 0,
 													  "modified": [">=", add_days(today(), -7)]
 												  },
-												  fields=["name", "prefix", "company",
-														  "last_communication_error"])
+												  fields=["name", "prefix", "company"])
 
 		if failed_communications:
 			frappe.logger().warning(
-				f"⚠️ Found {len(failed_communications)} failed communications in last 7 days")
-
-			# Tentar re-comunicar automaticamente
-			for series in failed_communications:
-				try_auto_retry_communication(series)
+				f"⚠️ Found {len(failed_communications)} uncommunicated series in last 7 days")
 
 		# Séries não comunicadas há muito tempo
 		overdue_series = frappe.db.get_all("Portugal Series Configuration",
 										   filters={
 											   "is_communicated": 0,
-											   "enable_at_communication": 1,
 											   "is_active": 1,
 											   "creation": ["<", add_days(today(), -3)]
 										   },
@@ -442,39 +433,19 @@ def monitor_communication_failures():
 		frappe.log_error(f"Error monitoring communication failures: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Tentar re-comunicação automática
-def try_auto_retry_communication(series):
-	"""
-	Tenta re-comunicar série automaticamente
-	"""
-	try:
-		# Verificar se tem credenciais
-		if series.get("at_username") and series.get("at_password"):
-			frappe.logger().info(f"🔄 Attempting auto-retry for series {series.name}")
-
-			# Chamar método de comunicação
-			frappe.call(
-				"portugal_compliance.doctype.portugal_series_configuration.portugal_series_configuration.register_series_at",
-				**{"doc": series.name})
-
-	except Exception as e:
-		frappe.log_error(f"Error in auto-retry communication for {series.name}: {str(e)}")
-
-
-# ✅ NOVA FUNÇÃO: Atualizar tendências de séries
 def update_series_trends():
 	"""
-	Atualiza tendências de uso das séries
+	Atualiza tendências de uso das séries - ADAPTADA
 	"""
 	try:
 		active_series = frappe.db.get_all("Portugal Series Configuration",
 										  filters={"is_active": 1},
-										  fields=["name"])
+										  fields=["name", "prefix", "document_type"])
 
 		for series in active_series:
 			try:
-				# Calcular tendência de uso
-				trend = calculate_series_usage_trend(series.name)
+				# ✅ ADAPTADO: Calcular tendência usando naming_series
+				trend = calculate_series_usage_trend_new_approach(series)
 
 				# Atualizar no documento
 				frappe.db.set_value("Portugal Series Configuration", series.name, {
@@ -489,24 +460,23 @@ def update_series_trends():
 		frappe.log_error(f"Error updating series trends: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Calcular tendência de uso
-def calculate_series_usage_trend(series_name):
+# ✅ FUNÇÃO ADAPTADA: Calcular tendência usando naming_series
+def calculate_series_usage_trend_new_approach(series):
 	"""
-	Calcula tendência de uso de uma série
+	Calcula tendência de uso usando naming_series nativa
 	"""
 	try:
-		# Obter série
-		series_doc = frappe.get_doc("Portugal Series Configuration", series_name)
+		naming_series = f"{series.prefix}.####"
 
 		# Contar documentos dos últimos 30 dias
-		recent_docs = frappe.db.count(series_doc.document_type, {
-			"portugal_series": series_name,
+		recent_docs = frappe.db.count(series.document_type, {
+			"naming_series": naming_series,
 			"creation": [">=", add_days(today(), -30)]
 		})
 
 		# Contar documentos dos 30 dias anteriores
-		previous_docs = frappe.db.count(series_doc.document_type, {
-			"portugal_series": series_name,
+		previous_docs = frappe.db.count(series.document_type, {
+			"naming_series": naming_series,
 			"creation": ["between", [add_days(today(), -60), add_days(today(), -30)]]
 		})
 
@@ -529,37 +499,29 @@ def calculate_series_usage_trend(series_name):
 		return "Unknown"
 
 
-# ✅ NOVA FUNÇÃO: Verificar consistência das naming series
 def check_naming_series_consistency():
 	"""
-	Verifica consistência das naming series com as séries portuguesas
+	Verifica consistência das naming series - ADAPTADA
 	"""
 	try:
 		inconsistent_series = []
 
 		active_series = frappe.db.get_all("Portugal Series Configuration",
 										  filters={"is_active": 1},
-										  fields=["name", "prefix", "document_type"])
+										  fields=["name", "prefix", "document_type",
+												  "naming_series"])
 
 		for series in active_series:
 			expected_naming = f"{series.prefix}.####"
 
-			try:
-				# Verificar se naming series está no DocType
-				doctype_meta = frappe.get_meta(series.document_type)
-				autoname_options = getattr(doctype_meta, 'autoname', '') or ''
-
-				if expected_naming not in autoname_options:
-					inconsistent_series.append({
-						"series": series.name,
-						"prefix": series.prefix,
-						"document_type": series.document_type,
-						"expected": expected_naming,
-						"current": autoname_options
-					})
-
-			except Exception as e:
-				frappe.log_error(f"Error checking naming series for {series.name}: {str(e)}")
+			if not series.naming_series or series.naming_series != expected_naming:
+				inconsistent_series.append({
+					"series": series.name,
+					"prefix": series.prefix,
+					"document_type": series.document_type,
+					"expected": expected_naming,
+					"current": series.naming_series or "Not Set"
+				})
 
 		if inconsistent_series:
 			frappe.logger().warning(
@@ -567,44 +529,41 @@ def check_naming_series_consistency():
 
 			# Tentar corrigir automaticamente
 			for series_info in inconsistent_series:
-				try_fix_naming_series(series_info)
+				try_fix_naming_series_configuration(series_info)
 
 	except Exception as e:
 		frappe.log_error(f"Error checking naming series consistency: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Tentar corrigir naming series
-def try_fix_naming_series(series_info):
+# ✅ FUNÇÃO ADAPTADA: Corrigir naming series configuration
+def try_fix_naming_series_configuration(series_info):
 	"""
-	Tenta corrigir naming series automaticamente
+	Tenta corrigir naming series na configuração
 	"""
 	try:
-		doctype = series_info["document_type"]
 		expected_naming = series_info["expected"]
-		current_options = series_info["current"]
 
-		# Adicionar naming series se não existir
-		if expected_naming not in current_options:
-			new_options = current_options + f"\n{expected_naming}" if current_options else expected_naming
+		# Atualizar Portugal Series Configuration
+		frappe.db.set_value("Portugal Series Configuration", series_info["series"],
+							"naming_series", expected_naming)
 
-			frappe.db.set_value("DocType", doctype, "autoname", new_options)
-			frappe.clear_cache(doctype=doctype)
-
-			frappe.logger().info(f"✅ Fixed naming series for {doctype}: added {expected_naming}")
+		frappe.logger().info(
+			f"✅ Fixed naming series for {series_info['series']}: {expected_naming}")
 
 	except Exception as e:
-		frappe.log_error(f"Error fixing naming series: {str(e)}")
+		frappe.log_error(f"Error fixing naming series configuration: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Validar integridade dos campos customizados
-def validate_custom_fields_integrity():
+# ✅ FUNÇÃO ADAPTADA: Validar apenas campos essenciais
+def validate_essential_custom_fields_integrity():
 	"""
-	Valida se campos customizados estão presentes em todos os DocTypes necessários
+	Valida se campos customizados ESSENCIAIS estão presentes (SEM portugal_series)
 	"""
 	try:
-		required_fields = ["atcud_code", "portugal_series"]
+		# ✅ ADAPTADO: Apenas campos essenciais (removido portugal_series)
+		required_fields = ["atcud_code"]
 		doctypes_to_check = [
-			"Sales Invoice", "Purchase Invoice", "Payment Entry",
+			"Sales Invoice", "POS Invoice", "Purchase Invoice", "Payment Entry",
 			"Delivery Note", "Purchase Receipt", "Journal Entry", "Stock Entry"
 		]
 
@@ -624,27 +583,29 @@ def validate_custom_fields_integrity():
 					})
 
 		if missing_fields:
-			frappe.logger().warning(f"⚠️ Found {len(missing_fields)} missing custom fields")
+			frappe.logger().warning(
+				f"⚠️ Found {len(missing_fields)} missing essential custom fields")
 
 			# Tentar criar campos em falta
 			for missing in missing_fields:
-				try_create_missing_field(missing)
+				try_create_essential_field(missing)
 		else:
-			frappe.logger().info("✅ All custom fields are present")
+			frappe.logger().info("✅ All essential custom fields are present")
 
 	except Exception as e:
-		frappe.log_error(f"Error validating custom fields integrity: {str(e)}")
+		frappe.log_error(f"Error validating essential custom fields integrity: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Tentar criar campo em falta
-def try_create_missing_field(missing_field):
+# ✅ FUNÇÃO ADAPTADA: Criar apenas campos essenciais
+def try_create_essential_field(missing_field):
 	"""
-	Tenta criar campo customizado em falta
+	Tenta criar campo customizado essencial (SEM portugal_series)
 	"""
 	try:
 		doctype = missing_field["doctype"]
 		field = missing_field["field"]
 
+		# ✅ ADAPTADO: Apenas criar atcud_code (removido portugal_series)
 		if field == "atcud_code":
 			field_doc = frappe.get_doc({
 				"doctype": "Custom Field",
@@ -658,46 +619,28 @@ def try_create_missing_field(missing_field):
 				"print_hide": 0,
 				"bold": 1,
 				"in_list_view": 1,
-				"description": "Código Único de Documento - obrigatório em Portugal"
+				"description": "Código Único de Documento - gerado automaticamente"
 			})
-		elif field == "portugal_series":
-			field_doc = frappe.get_doc({
-				"doctype": "Custom Field",
-				"dt": doctype,
-				"module": "Portugal Compliance",
-				"fieldname": "portugal_series",
-				"label": "Portugal Series",
-				"fieldtype": "Link",
-				"options": "Portugal Series Configuration",
-				"insert_after": "atcud_code",
-				"reqd": 1,
-				"in_list_view": 1,
-				"description": "Série portuguesa configurada para este documento"
-			})
-		else:
-			return
 
-		field_doc.insert(ignore_permissions=True)
-		frappe.logger().info(f"✅ Created missing field {field} for {doctype}")
+			field_doc.insert(ignore_permissions=True)
+			frappe.logger().info(f"✅ Created missing field {field} for {doctype}")
 
 	except Exception as e:
-		frappe.log_error(f"Error creating missing field {missing_field}: {str(e)}")
+		frappe.log_error(f"Error creating essential field {missing_field}: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Limpar comunicações falhadas
 def cleanup_failed_communications():
 	"""
-	Limpa dados de comunicações falhadas antigas
+	Limpa dados de comunicações falhadas antigas - MANTIDA
 	"""
 	try:
 		# Limpar erros de comunicação antigos (mais de 30 dias)
-		old_errors = frappe.db.sql("""
-								   UPDATE `tabPortugal Series Configuration`
-								   SET last_communication_error = NULL,
-									   communication_status     = NULL
-								   WHERE communication_status = 'Failed'
-									 AND modified < %s
-								   """, add_days(today(), -30))
+		frappe.db.sql("""
+					  UPDATE `tabPortugal Series Configuration`
+					  SET last_communication_error = NULL
+					  WHERE modified < %s
+						AND is_communicated = 0
+					  """, add_days(today(), -30))
 
 		frappe.logger().info("🧹 Cleaned up old communication errors")
 
@@ -705,10 +648,9 @@ def cleanup_failed_communications():
 		frappe.log_error(f"Error cleaning up failed communications: {str(e)}")
 
 
-# ✅ NOVA FUNÇÃO: Gerar métricas de compliance
 def generate_compliance_metrics():
 	"""
-	Gera métricas de compliance para dashboard
+	Gera métricas de compliance para dashboard - ADAPTADA
 	"""
 	try:
 		metrics = {
@@ -717,7 +659,7 @@ def generate_compliance_metrics():
 			"active_series": frappe.db.count("Portugal Series Configuration", {"is_active": 1}),
 			"communicated_series": frappe.db.count("Portugal Series Configuration",
 												   {"is_communicated": 1}),
-			"new_format_series": count_series_with_new_format(),
+			"correct_naming_series": count_series_with_correct_naming_series(),
 			"total_atcuds": frappe.db.count("ATCUD Log"),
 			"successful_atcuds": frappe.db.count("ATCUD Log", {"generation_status": "Success"}),
 			"portuguese_companies": frappe.db.count("Company", {"country": "Portugal"}),
@@ -728,45 +670,118 @@ def generate_compliance_metrics():
 		# Armazenar no cache
 		frappe.cache.set("portugal_compliance_metrics", metrics, expires_in_sec=86400)
 
-		frappe.logger().info(f"📊 Generated compliance metrics: {metrics}")
+		frappe.logger().info(f"📊 Generated compliance metrics (New Approach): {metrics}")
 
 	except Exception as e:
 		frappe.log_error(f"Error generating compliance metrics: {str(e)}")
 
 
-# ========== FUNÇÕES DE ALERTA (NOVAS) ==========
-
-def create_prefix_format_alert(invalid_prefixes, legacy_prefixes):
+# ✅ NOVA FUNÇÃO: Contar séries com naming_series correto
+def count_series_with_correct_naming_series():
 	"""
-	Cria alerta para formatos de prefixo incorretos
+	Conta séries com naming_series configurado corretamente
+	"""
+	try:
+		correct_count = 0
+
+		all_series = frappe.db.get_all("Portugal Series Configuration",
+									   fields=["prefix", "naming_series"])
+
+		for series in all_series:
+			if series.prefix and series.naming_series:
+				expected = f"{series.prefix}.####"
+				if series.naming_series == expected:
+					correct_count += 1
+
+		return correct_count
+
+	except Exception:
+		return 0
+
+
+# ✅ NOVA FUNÇÃO: Sincronizar Portugal Series Configuration
+def sync_portugal_series_configurations():
+	"""
+	Sincroniza Portugal Series Configuration com Portugal Document Series
+	"""
+	try:
+		# Verificar se há Portugal Document Series para sincronizar
+		if frappe.db.table_exists("tabPortugal Document Series"):
+			document_series = frappe.db.get_all("Portugal Document Series",
+												filters={"is_active": 1},
+												fields=["name", "prefix", "document_type",
+														"company"])
+
+			synced_count = 0
+			for ds in document_series:
+				try:
+					# Verificar se já existe Portugal Series Configuration correspondente
+					existing_config = frappe.db.exists("Portugal Series Configuration", {
+						"prefix": ds.prefix,
+						"company": ds.company,
+						"document_type": ds.document_type
+					})
+
+					if not existing_config:
+						# Criar Portugal Series Configuration
+						config = frappe.get_doc({
+							"doctype": "Portugal Series Configuration",
+							"series_name": f"{ds.document_type} - {ds.prefix}",
+							"company": ds.company,
+							"document_type": ds.document_type,
+							"prefix": ds.prefix,
+							"naming_series": f"{ds.prefix}.####",
+							"current_sequence": 1,
+							"is_active": 1,
+							"is_communicated": 0
+						})
+						config.insert(ignore_permissions=True)
+						synced_count += 1
+
+				except Exception as e:
+					frappe.log_error(f"Error syncing series {ds.name}: {str(e)}")
+
+			if synced_count > 0:
+				frappe.logger().info(
+					f"✅ Synced {synced_count} Portugal Document Series to Portugal Series Configuration")
+
+	except Exception as e:
+		frappe.log_error(f"Error syncing Portugal Series Configurations: {str(e)}")
+
+
+# ========== FUNÇÕES DE ALERTA ADAPTADAS ==========
+
+def create_naming_series_format_alert(invalid_naming_series, missing_naming_series):
+	"""
+	Cria alerta para formatos de naming series incorretos
 	"""
 	try:
 		admin_users = get_compliance_users()
 
 		message = f"""
-		⚠️ Alerta de Formato de Prefixo:
-		- {len(invalid_prefixes)} séries com formato inválido
-		- {len(legacy_prefixes)} séries com formato legado
+		⚠️ Alerta de Naming Series:
+		- {len(invalid_naming_series)} séries com naming series inválido
+		- {len(missing_naming_series)} séries sem naming series
 
-		Recomenda-se corrigir os formatos para XX-YYYY-COMPANY
+		Recomenda-se corrigir para formato: PREFIX.####
 		"""
 
 		for user in admin_users:
 			frappe.get_doc({
 				"doctype": "Notification Log",
-				"subject": "Portugal Compliance: Formato de Prefixo",
+				"subject": "Portugal Compliance: Naming Series",
 				"email_content": message,
 				"for_user": user,
 				"type": "Alert"
 			}).insert(ignore_permissions=True)
 
 	except Exception as e:
-		frappe.log_error(f"Error creating prefix format alert: {str(e)}")
+		frappe.log_error(f"Error creating naming series format alert: {str(e)}")
 
 
 def create_atcud_integrity_alert(duplicate_atcuds):
 	"""
-	Cria alerta crítico para duplicação de ATCUD
+	Cria alerta crítico para duplicação de ATCUD - MANTIDA
 	"""
 	try:
 		admin_users = get_compliance_users()
@@ -792,16 +807,14 @@ def create_atcud_integrity_alert(duplicate_atcuds):
 		frappe.log_error(f"Error creating ATCUD integrity alert: {str(e)}")
 
 
-# ========== MANTER TODAS AS FUNÇÕES EXISTENTES ==========
-# (check_series_expiration, create_expiry_notification, validate_daily_sequences,
-#  cleanup_old_logs, check_pending_communications, etc.)
+# ========== MANTER TODAS AS FUNÇÕES EXISTENTES ADAPTADAS ==========
 
 def check_series_expiration():
 	"""
-	Verifica séries que estão a expirar - MANTIDA
+	Verifica séries que estão a expirar - ADAPTADA
 	"""
 	try:
-		# Séries que expiram nos próximos 30 dias
+		# ✅ ADAPTADO: Usar Portugal Series Configuration
 		expiring_soon = frappe.db.sql("""
 									  SELECT name, series_name, company, expiry_date
 									  FROM `tabPortugal Series Configuration`
@@ -879,7 +892,7 @@ def validate_daily_sequences():
 
 def check_sequence_gaps():
 	"""
-	Verifica gaps nas sequências de documentos - MANTIDA
+	Verifica gaps nas sequências de documentos - ADAPTADA
 	"""
 	try:
 		active_series = frappe.db.get_all("Portugal Series Configuration",
@@ -907,7 +920,7 @@ def find_sequence_gaps(series_prefix):
 
 def cleanup_old_logs():
 	"""
-	Limpa logs antigos - MELHORADA
+	Limpa logs antigos - MANTIDA
 	"""
 	try:
 		# Manter logs por 90 dias
@@ -938,7 +951,7 @@ def cleanup_old_logs():
 
 def check_pending_communications():
 	"""
-	Verifica comunicações pendentes com a AT - MANTIDA
+	Verifica comunicações pendentes com a AT - ADAPTADA
 	"""
 	try:
 		pending_series = frappe.db.sql("""
@@ -961,13 +974,14 @@ def check_pending_communications():
 
 def try_auto_communication(series):
 	"""
-	Tenta comunicar série automaticamente - MANTIDA
+	Tenta comunicar série automaticamente - ADAPTADA
 	"""
 	try:
-		auth_settings = frappe.db.get_value("Portugal Auth Settings",
-											series.company, ["username", "password"], as_dict=True)
+		# ✅ ADAPTADO: Verificar credenciais na empresa
+		company_doc = frappe.get_doc("Company", series.company)
 
-		if auth_settings and auth_settings.username:
+		if (getattr(company_doc, 'at_username', None) and
+			getattr(company_doc, 'portugal_compliance_enabled', 0)):
 			frappe.logger().info(f"Attempting auto-communication for series {series.name}")
 
 	except Exception as e:
@@ -976,7 +990,7 @@ def try_auto_communication(series):
 
 def update_usage_statistics():
 	"""
-	Atualiza estatísticas de uso - MANTIDA
+	Atualiza estatísticas de uso - ADAPTADA
 	"""
 	try:
 		active_series = frappe.db.get_all("Portugal Series Configuration",
@@ -992,12 +1006,26 @@ def update_usage_statistics():
 
 def update_series_statistics(series_name):
 	"""
-	Atualiza estatísticas de uma série específica - MANTIDA
+	Atualiza estatísticas de uma série específica - ADAPTADA
 	"""
 	try:
 		series_doc = frappe.get_doc("Portugal Series Configuration", series_name)
-		if hasattr(series_doc, 'update_usage_statistics'):
-			series_doc.update_usage_statistics()
+
+		# ✅ ADAPTADO: Calcular estatísticas usando naming_series
+		naming_series = f"{series_doc.prefix}.####"
+
+		# Contar documentos que usam esta naming_series
+		total_docs = frappe.db.count(series_doc.document_type, {
+			"naming_series": naming_series,
+			"company": series_doc.company,
+			"docstatus": ["!=", 2]
+		})
+
+		# Atualizar estatísticas
+		frappe.db.set_value("Portugal Series Configuration", series_name, {
+			"total_documents_issued": total_docs,
+			"last_statistics_update": now()
+		}, update_modified=False)
 
 	except Exception as e:
 		frappe.log_error(f"Error updating statistics for series {series_name}: {str(e)}")
@@ -1005,14 +1033,15 @@ def update_series_statistics(series_name):
 
 def backup_critical_data():
 	"""
-	Faz backup de dados críticos - MANTIDA
+	Faz backup de dados críticos - ADAPTADA
 	"""
 	try:
 		backup_data = {
 			"date": today(),
 			"series_count": frappe.db.count("Portugal Series Configuration"),
 			"atcud_count": frappe.db.count("ATCUD Log"),
-			"companies_count": frappe.db.count("Company", {"portugal_compliance_enabled": 1})
+			"companies_count": frappe.db.count("Company", {"portugal_compliance_enabled": 1}),
+			"naming_series_health": check_naming_series_health()
 		}
 
 		frappe.cache.set("portugal_compliance_daily_backup", backup_data, expires_in_sec=604800)
@@ -1036,11 +1065,12 @@ def send_daily_notifications():
 
 def check_critical_alerts():
 	"""
-	Verifica alertas críticos - MANTIDA
+	Verifica alertas críticos - ADAPTADA
 	"""
 	try:
 		alerts = []
 
+		# ✅ ADAPTADO: Usar Portugal Series Configuration
 		overdue_series = frappe.db.count("Portugal Series Configuration", {
 			"is_communicated": 0,
 			"is_active": 1,
@@ -1086,10 +1116,9 @@ def send_alert_notifications(alerts):
 
 def get_compliance_users(company=None):
 	"""
-	Obtém utilizadores responsáveis pelo compliance - MELHORADA
+	Obtém utilizadores responsáveis pelo compliance - MANTIDA
 	"""
 	try:
-		# ✅ MELHORADA: Incluir mais roles relevantes
 		users = frappe.db.sql("""
 							  SELECT DISTINCT u.name
 							  FROM `tabUser` u
